refactor: introduce interface and hide implementation details
